vana class 2: dead -1 sentinels in quaerere and sententia_fissio
quaerere's backward pass wrapped its unsigned index past zero and
read out of bounds (dead < 0 break, now s32 and live). sententia's
_invenire_finem_sententiae loses its -1 sentinel entirely - it
returns textus.mensura, which is what every caller clamped -1 into
through checks that could never fire.

+++ lib/quaerere.c
@@ -288,7 +288,7 @@ _quaerere_passus_posterior(
     b32    case_sensitivus)
 {
     i32 qi;
-    i32 si_idx;
+    s32 si_idx;
 
     si (quaestio.mensura <= 1)
     {
@@ -297,7 +297,7 @@ _quaerere_passus_posterior(
 
     /* Start from last match position, scan backward */
     qi = quaestio.mensura - 1;
-    si_idx = positiones[qi];
+    si_idx = (s32)positiones[qi];
 
     dum (qi > 0)
     {
@@ -313,9 +313,9 @@ _quaerere_passus_posterior(
                                     case_sensitivus))
         {
             /* Check if this position is better (tighter clustering) */
-            si (si_idx > positiones[qi - 1])
+            si (si_idx > (s32)positiones[qi - 1])
             {
-                positiones[qi - 1] = si_idx;
+                positiones[qi - 1] = (i32)si_idx;
             }
             qi--;
         }
